requirements.txt
# LangChain核心 - 使用兼容版本
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-text-splitters>=0.3.8
langchain-core>=0.3.0

# 向量存储
faiss-cpu>=1.8.0

# PDF处理
pypdf>=5.0.0

# Streamlit前端
streamlit>=1.40.0

# 本地 Embedding（免费）
sentence-transformers>=2.2.0
langchain-huggingface>=0.0.1

# 其他工具
python-dotenv>=1.0.0
pandas>=2.0.0
